Move wasm aminotypes to modules/wasm/aminomessages
