+++ packages/cosmwasm-stargate/src/modules/wasm/aminomessages.spec.ts
@@ -18,8 +18,8 @@ import {
   AminoMsgMigrateContract,
   AminoMsgStoreCode,
   AminoMsgUpdateAdmin,
-  cosmWasmTypes,
-} from "./aminotypes";
+  createWasmAminoConverters,
+} from "./aminomessages";
 
 describe("AminoTypes", () => {
   describe("toAmino", () => {
@@ -29,7 +29,7 @@ describe("AminoTypes", () => {
         wasmByteCode: fromBase64("WUVMTE9XIFNVQk1BUklORQ=="),
         instantiatePermission: undefined,
       };
-      const aminoMsg = new AminoTypes({ prefix: "cosmos", additions: cosmWasmTypes }).toAmino({
+      const aminoMsg = new AminoTypes({ prefix: "cosmos", additions: createWasmAminoConverters() }).toAmino({
         typeUrl: "/cosmwasm.wasm.v1.MsgStoreCode",
         value: msg,
       });
@@ -54,10 +54,12 @@ describe("AminoTypes", () => {
           funds: coins(1234, "ucosm"),
           admin: "cosmos10dyr9899g6t0pelew4nvf4j5c3jcgv0r73qga5",
         };
-        const aminoMsg = new AminoTypes({ prefix: "cosmos", additions: cosmWasmTypes }).toAmino({
-          typeUrl: "/cosmwasm.wasm.v1.MsgInstantiateContract",
-          value: msg,
-        });
+        const aminoMsg = new AminoTypes({ prefix: "cosmos", additions: createWasmAminoConverters() }).toAmino(
+          {
+            typeUrl: "/cosmwasm.wasm.v1.MsgInstantiateContract",
+            value: msg,
+          },
+        );
         const expected: AminoMsgInstantiateContract = {
           type: "wasm/MsgInstantiateContract",
           value: {
@@ -82,10 +84,12 @@ describe("AminoTypes", () => {
           funds: coins(1234, "ucosm"),
           admin: "",
         };
-        const aminoMsg = new AminoTypes({ prefix: "cosmos", additions: cosmWasmTypes }).toAmino({
-          typeUrl: "/cosmwasm.wasm.v1.MsgInstantiateContract",
-          value: msg,
-        });
+        const aminoMsg = new AminoTypes({ prefix: "cosmos", additions: createWasmAminoConverters() }).toAmino(
+          {
+            typeUrl: "/cosmwasm.wasm.v1.MsgInstantiateContract",
+            value: msg,
+          },
+        );
         const expected: AminoMsgInstantiateContract = {
           type: "wasm/MsgInstantiateContract",
           value: {
@@ -107,7 +111,7 @@ describe("AminoTypes", () => {
         newAdmin: "cosmos10dyr9899g6t0pelew4nvf4j5c3jcgv0r73qga5",
         contract: "cosmos1xy4yqngt0nlkdcenxymg8tenrghmek4nmqm28k",
       };
-      const aminoMsg = new AminoTypes({ prefix: "cosmos", additions: cosmWasmTypes }).toAmino({
+      const aminoMsg = new AminoTypes({ prefix: "cosmos", additions: createWasmAminoConverters() }).toAmino({
         typeUrl: "/cosmwasm.wasm.v1.MsgUpdateAdmin",
         value: msg,
       });
@@ -127,7 +131,7 @@ describe("AminoTypes", () => {
         sender: "cosmos1pkptre7fdkl6gfrzlesjjvhxhlc3r4gmmk8rs6",
         contract: "cosmos1xy4yqngt0nlkdcenxymg8tenrghmek4nmqm28k",
       };
-      const aminoMsg = new AminoTypes({ prefix: "cosmos", additions: cosmWasmTypes }).toAmino({
+      const aminoMsg = new AminoTypes({ prefix: "cosmos", additions: createWasmAminoConverters() }).toAmino({
         typeUrl: "/cosmwasm.wasm.v1.MsgClearAdmin",
         value: msg,
       });
@@ -148,7 +152,7 @@ describe("AminoTypes", () => {
         msg: toUtf8(`{"foo":"bar"}`),
         funds: coins(1234, "ucosm"),
       };
-      const aminoMsg = new AminoTypes({ prefix: "cosmos", additions: cosmWasmTypes }).toAmino({
+      const aminoMsg = new AminoTypes({ prefix: "cosmos", additions: createWasmAminoConverters() }).toAmino({
         typeUrl: "/cosmwasm.wasm.v1.MsgExecuteContract",
         value: msg,
       });
@@ -171,7 +175,7 @@ describe("AminoTypes", () => {
         codeId: Long.fromString("98765"),
         msg: toUtf8(`{"foo":"bar"}`),
       };
-      const aminoMsg = new AminoTypes({ prefix: "cosmos", additions: cosmWasmTypes }).toAmino({
+      const aminoMsg = new AminoTypes({ prefix: "cosmos", additions: createWasmAminoConverters() }).toAmino({
         typeUrl: "/cosmwasm.wasm.v1.MsgMigrateContract",
         value: msg,
       });
@@ -197,7 +201,9 @@ describe("AminoTypes", () => {
           wasm_byte_code: "WUVMTE9XIFNVQk1BUklORQ==",
         },
       };
-      const msg = new AminoTypes({ prefix: "cosmos", additions: cosmWasmTypes }).fromAmino(aminoMsg);
+      const msg = new AminoTypes({ prefix: "cosmos", additions: createWasmAminoConverters() }).fromAmino(
+        aminoMsg,
+      );
       const expectedValue: MsgStoreCode = {
         sender: "cosmos1pkptre7fdkl6gfrzlesjjvhxhlc3r4gmmk8rs6",
         wasmByteCode: fromBase64("WUVMTE9XIFNVQk1BUklORQ=="),
@@ -223,7 +229,9 @@ describe("AminoTypes", () => {
             admin: "cosmos10dyr9899g6t0pelew4nvf4j5c3jcgv0r73qga5",
           },
         };
-        const msg = new AminoTypes({ prefix: "cosmos", additions: cosmWasmTypes }).fromAmino(aminoMsg);
+        const msg = new AminoTypes({ prefix: "cosmos", additions: createWasmAminoConverters() }).fromAmino(
+          aminoMsg,
+        );
         const expectedValue: MsgInstantiateContract = {
           sender: "cosmos1pkptre7fdkl6gfrzlesjjvhxhlc3r4gmmk8rs6",
           codeId: Long.fromString("12345"),
@@ -250,7 +258,9 @@ describe("AminoTypes", () => {
             funds: coins(1234, "ucosm"),
           },
         };
-        const msg = new AminoTypes({ prefix: "cosmos", additions: cosmWasmTypes }).fromAmino(aminoMsg);
+        const msg = new AminoTypes({ prefix: "cosmos", additions: createWasmAminoConverters() }).fromAmino(
+          aminoMsg,
+        );
         const expectedValue: MsgInstantiateContract = {
           sender: "cosmos1pkptre7fdkl6gfrzlesjjvhxhlc3r4gmmk8rs6",
           codeId: Long.fromString("12345"),
@@ -275,7 +285,9 @@ describe("AminoTypes", () => {
           contract: "cosmos1xy4yqngt0nlkdcenxymg8tenrghmek4nmqm28k",
         },
       };
-      const msg = new AminoTypes({ prefix: "cosmos", additions: cosmWasmTypes }).fromAmino(aminoMsg);
+      const msg = new AminoTypes({ prefix: "cosmos", additions: createWasmAminoConverters() }).fromAmino(
+        aminoMsg,
+      );
       const expectedValue: MsgUpdateAdmin = {
         sender: "cosmos1pkptre7fdkl6gfrzlesjjvhxhlc3r4gmmk8rs6",
         newAdmin: "cosmos10dyr9899g6t0pelew4nvf4j5c3jcgv0r73qga5",
@@ -295,7 +307,9 @@ describe("AminoTypes", () => {
           contract: "cosmos1xy4yqngt0nlkdcenxymg8tenrghmek4nmqm28k",
         },
       };
-      const msg = new AminoTypes({ prefix: "cosmos", additions: cosmWasmTypes }).fromAmino(aminoMsg);
+      const msg = new AminoTypes({ prefix: "cosmos", additions: createWasmAminoConverters() }).fromAmino(
+        aminoMsg,
+      );
       const expectedValue: MsgClearAdmin = {
         sender: "cosmos1pkptre7fdkl6gfrzlesjjvhxhlc3r4gmmk8rs6",
         contract: "cosmos1xy4yqngt0nlkdcenxymg8tenrghmek4nmqm28k",
@@ -316,7 +330,9 @@ describe("AminoTypes", () => {
           funds: coins(1234, "ucosm"),
         },
       };
-      const msg = new AminoTypes({ prefix: "cosmos", additions: cosmWasmTypes }).fromAmino(aminoMsg);
+      const msg = new AminoTypes({ prefix: "cosmos", additions: createWasmAminoConverters() }).fromAmino(
+        aminoMsg,
+      );
       const expectedValue: MsgExecuteContract = {
         sender: "cosmos1pkptre7fdkl6gfrzlesjjvhxhlc3r4gmmk8rs6",
         contract: "cosmos1xy4yqngt0nlkdcenxymg8tenrghmek4nmqm28k",
@@ -339,7 +355,9 @@ describe("AminoTypes", () => {
           msg: { foo: "bar" },
         },
       };
-      const msg = new AminoTypes({ prefix: "cosmos", additions: cosmWasmTypes }).fromAmino(aminoMsg);
+      const msg = new AminoTypes({ prefix: "cosmos", additions: createWasmAminoConverters() }).fromAmino(
+        aminoMsg,
+      );
       const expectedValue: MsgMigrateContract = {
         sender: "cosmos1pkptre7fdkl6gfrzlesjjvhxhlc3r4gmmk8rs6",
         contract: "cosmos1xy4yqngt0nlkdcenxymg8tenrghmek4nmqm28k",
